Switching to the new simple type iimplementation. This enable us to get more information from simple types, which is essential to PSVI support. 1. Modified xs package to use the new simple type interfaces. 2. Fixed bugs in the new simple type implementation.
git-svn-id: https://svn.apache.org/repos/asf/xerces/java/trunk@317956 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/xerces/impl/dv/xs_new/MonthDV.java
@@ -141,10 +141,10 @@ public class MonthDV extends AbstractDateTimeDV {
      * @param date2
      * @return
      */
-    protected  int compareDates(int[] date1, int[] date2) {
+    protected  short compareDates(int[] date1, int[] date2) {
 
         if ( date1[utc]==date2[utc] ) {
-            return (date1[M]>=date2[M])?(date1[M]>date2[M])?1:0:-1;
+            return (short)((date1[M]>=date2[M])?(date1[M]>date2[M])?1:0:-1);
         }
 
         if ( date1[utc]=='Z' || date2[utc]=='Z' ) {
